+++ assets-proxy/src/main/java/es/nttdata/assetsproxy/infrastructure/apirest/dto/Asset.java
@@ -7,6 +7,6 @@ public record AssetResponseDto(
         String filename,
         String contentType,
         String url,
-        Long size,
+        Integer size,
         OffsetDateTime uploadDate
 ) { }
